[FLINK-18656][tests] Rename MultipleInputStreamTaskTestHarnessBuilder to StreamTaskMailboxTestHarnessBuilder
The orignal concept MultipleInputStreamTaskTestHarnessBuilder proved much more
versatile then initially expected and it can easily handle all of the uses cases:
- MultipleInputStreamTask
- OneInputStreamTask
- SourceStreamTask

Hence there is no need for the abstraction and no need to provide specialized versions of
MultipleInputStreamTaskTestHarnessBuilder for the other types of tasks.
